+++ lib/WorseReflection/Bridge/TolerantParser/Diagnostics/MissingDocblockProvider.php
@@ -2,7 +2,6 @@
 
 namespace Phpactor\WorseReflection\Bridge\TolerantParser\Diagnostics;
 
-use Generator;
 use Microsoft\PhpParser\Node;
 use Microsoft\PhpParser\Node\MethodDeclaration;
 use Phpactor\WorseReflection\Core\DiagnosticProvider;
@@ -15,7 +14,7 @@ use Phpactor\WorseReflection\Core\Util\NodeUtil;
 
 class MissingDocblockProvider implements DiagnosticProvider
 {
-    public function provide(NodeContextResolver $resolver, Frame $frame, Node $node): Generator
+    public function exit(NodeContextResolver $resolver, Frame $frame, Node $node): iterable
     {
         if (!$node instanceof MethodDeclaration) {
             return;
@@ -111,5 +110,10 @@ class MissingDocblockProvider implements DiagnosticProvider
             $methodName,
             $actualReturnType->__toString(),
         );
+    }
+
+    public function enter(NodeContextResolver $resolver, Frame $frame, Node $node): iterable
+    {
+        return [];
     }
 }