+++ lib/WorseReflection/Bridge/TolerantParser/Diagnostics/MissingReturnTypeProvider.php
@@ -2,7 +2,6 @@
 
 namespace Phpactor\WorseReflection\Bridge\TolerantParser\Diagnostics;
 
-use Generator;
 use Microsoft\PhpParser\Node;
 use Microsoft\PhpParser\Node\MethodDeclaration;
 use Phpactor\WorseReflection\Core\DiagnosticProvider;
@@ -14,7 +13,7 @@ use Phpactor\WorseReflection\Core\Util\NodeUtil;
 
 class MissingReturnTypeProvider implements DiagnosticProvider
 {
-    public function provide(NodeContextResolver $resolver, Frame $frame, Node $node): Generator
+    public function exit(NodeContextResolver $resolver, Frame $frame, Node $node): iterable
     {
         if (!$node instanceof MethodDeclaration) {
             return;
@@ -83,5 +82,10 @@ class MissingReturnTypeProvider implements DiagnosticProvider
             $methodName,
             $returnType->generalize()->reduce()
         );
+    }
+
+    public function enter(NodeContextResolver $resolver, Frame $frame, Node $node): iterable
+    {
+        return [];
     }
 }