+++ lib/WorseReflection/Core/Inference/NodeContextResolver.php
@@ -6,6 +6,7 @@ use Microsoft\PhpParser\MissingToken;
 use Microsoft\PhpParser\Node;
 use Microsoft\PhpParser\Token;
 use Phpactor\WorseReflection\Core\Cache;
+use Phpactor\WorseReflection\Core\DocBlock\DocBlockFactory;
 use Phpactor\WorseReflection\Core\Exception\CouldNotResolveNode;
 use Phpactor\WorseReflection\Core\Name;
 use Phpactor\WorseReflection\Reflector;
@@ -24,12 +25,15 @@ class NodeContextResolver
     private array $resolverMap;
 
     private Cache $cache;
+
+    private DocBlockFactory $docblockFactory;
     
     /**
      * @param array<class-name,Resolver> $resolverMap
      */
     public function __construct(
         Reflector $reflector,
+        DocBlockFactory $docblockFactory,
         LoggerInterface $logger,
         Cache $cache,
         array $resolverMap = []
@@ -38,11 +42,12 @@ class NodeContextResolver
         $this->reflector = $reflector;
         $this->resolverMap = $resolverMap;
         $this->cache = $cache;
+        $this->docblockFactory = $docblockFactory;
     }
 
     public function withCache(Cache $cache):self
     {
-        return new self($this->reflector, $this->logger, $cache, $this->resolverMap);
+        return new self($this->reflector, $this->docblockFactory, $this->logger, $cache, $this->resolverMap);
     }
 
     /**
@@ -61,6 +66,11 @@ class NodeContextResolver
     public function reflector(): Reflector
     {
         return $this->reflector;
+    }
+
+    public function docblockFactory(): DocBlockFactory
+    {
+        return $this->docblockFactory;
     }
 
     /**